+++ test/fixture/models.py
@@ -1,0 +1,11 @@
+from sqlalchemy import *
+
+# test rundiffs in shell
+meta_old_rundiffs = MetaData()
+meta_rundiffs = MetaData()
+
+tmp_account_rundiffs = Table('tmp_account_rundiffs', meta_rundiffs,
+    Column('id', Integer, primary_key=True),
+    Column('login', String(40)),
+    Column('passwd', String(40)),
+)