+++ test/fixture/shell.py
@@ -6,51 +6,22 @@ import shutil
 import sys
 import types
 
+from scripttest import TestFileEnvironment
+
 from test.fixture.pathed import *
 
 
 class Shell(Pathed):
     """Base class for command line tests"""
-    def execute(self, command, *p, **k):
-        """Return the fd of a command; can get output (stdout/err) and exitcode"""
-        # We might be passed a file descriptor for some reason; if so, just return it
-        if isinstance(command, types.FileType):
-            return command
 
-        # Redirect stderr to stdout
-        # This is a bit of a hack, but I've not found a better way
-        py_path = os.environ.get('PYTHONPATH', '')
-        py_path_list = py_path.split(':')
-        py_path_list.append(os.path.abspath('.'))
-        os.environ['PYTHONPATH'] = ':'.join(py_path_list)
-        fd = os.popen(command + ' 2>&1')
+    def setUp(self):
+        super(Shell, self).setUp()
+        self.env = TestFileEnvironment(os.path.join(self.temp_usable_dir, 'env'))
 
-        if py_path:
-            py_path = os.environ['PYTHONPATH'] = py_path
-        else:
-            del os.environ['PYTHONPATH']
-        return fd
+    def run_version(self, repos_path):
+        result = self.env.run('migrate version %s' % repos_path)
+        return int(result.stdout.strip())
 
-    def output_and_exitcode(self, *p, **k):
-        fd=self.execute(*p, **k)
-        output = fd.read().strip()
-        exitcode = fd.close()
-        if k.pop('emit',False):
-            print output
-        return (output, exitcode)
-
-    def exitcode(self, *p, **k):
-        """Execute a command and return its exit code
-        ...without printing its output/errors
-        """
-        ret = self.output_and_exitcode(*p, **k)
-        return ret[1]
-
-    def assertFailure(self, *p, **k):
-        output,exitcode = self.output_and_exitcode(*p, **k)
-        assert (exitcode), output
-
-    def assertSuccess(self, *p, **k):
-        output,exitcode = self.output_and_exitcode(*p, **k)
-        #self.assert_(not exitcode, output)
-        assert (not exitcode), output
+    def run_db_version(self, url, repos_path):
+        result = self.env.run('migrate db_version %s %s' % (url, repos_path))
+        return int(result.stdout.strip())